Joy_chatbot/requirements.txt
streamlit==1.32.0
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
bitsandbytes>=0.42.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
numpy>=1.26.0
pickle5
